requirements.txt
pillow
vosk
pyaudio
pyttsx3
numpy
nltk
keras
Pillow
